requirements.txt
asgiref==3.2.10
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cssselect2==0.3.0
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0
django-heroku==0.3.1
django-import-export==2.3.0
django-weasyprint==1.0.1
et-xmlfile==1.0.1
gunicorn==20.0.4
html5lib==1.1
httplib2==0.18.1
idna==2.10
jdcal==1.4.1
MarkupPy==1.14
numpy==1.19.1
odfpy==1.4.1
openpyxl==3.0.4
Pillow==7.2.0
psycopg2==2.8.5
pycparser==2.20
PyPDF2==1.26.0
Pyphen==0.9.5
python-decouple==3.3
pytz==2020.1
PyYAML==5.3.1
reportlab==3.5.49
requests==2.24.0
scipy==1.5.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
static3==0.7.0
tablib==2.0.0
tinycss2==1.0.2
urllib3==1.25.10
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.4
xlrd==1.2.0
XlsxWriter==1.3.6
xlwt==1.3.0
